Moved shard iterator specific impl to helpers
