+++ src/kinesis.rs
@@ -3,8 +3,7 @@ use std::fmt::Debug;
 use crate::kinesis::models::*;
 use async_trait::async_trait;
 use aws_sdk_kinesis::operation::get_shard_iterator::GetShardIteratorOutput;
-use aws_sdk_kinesis::{Client, Error};
-use chrono::Utc;
+use aws_sdk_kinesis::Error;
 use log::{debug, error};
 use tokio::sync::mpsc;
 use tokio::sync::mpsc::Sender;
@@ -13,61 +12,11 @@ use tokio::time::{sleep, Duration};
 pub mod helpers;
 pub mod models;
 
-pub fn new(
-    client: Client,
-    stream: String,
-    shard_id: String,
-    from_datetime: Option<chrono::DateTime<Utc>>,
-    tx_records: Sender<Result<ShardProcessorADT, PanicError>>,
-) -> Box<dyn ShardProcessor + Send + Sync> {
-    match from_datetime {
-        Some(from_datetime) => Box::new(ShardProcessorAtTimestamp {
-            config: ShardProcessorConfig {
-                client,
-                stream,
-                shard_id,
-                tx_records,
-            },
-            from_datetime,
-        }),
-        None => Box::new(ShardProcessorLatest {
-            config: ShardProcessorConfig {
-                client,
-                stream,
-                shard_id,
-                tx_records,
-            },
-        }),
-    }
-}
-
 #[async_trait]
 pub trait IteratorProvider: Send + Sync + Debug + Clone {
     fn get_config(&self) -> ShardProcessorConfig;
 
     async fn get_iterator(&self) -> Result<GetShardIteratorOutput, Error>;
-}
-
-#[async_trait]
-impl IteratorProvider for ShardProcessorLatest {
-    fn get_config(&self) -> ShardProcessorConfig {
-        self.config.clone()
-    }
-
-    async fn get_iterator(&self) -> Result<GetShardIteratorOutput, Error> {
-        helpers::get_latest_iterator(self.clone()).await
-    }
-}
-
-#[async_trait]
-impl IteratorProvider for ShardProcessorAtTimestamp {
-    fn get_config(&self) -> ShardProcessorConfig {
-        self.config.clone()
-    }
-
-    async fn get_iterator(&self) -> Result<GetShardIteratorOutput, Error> {
-        helpers::get_iterator_at_timestamp(self.clone(), self.from_datetime).await
-    }
 }
 
 #[async_trait]
@@ -233,3 +182,6 @@ where
         Ok(())
     }
 }
+
+#[cfg(test)]
+mod tests;
